Add Query implementation to Iot Hub service client
