+++ sdk/iot/Azure.Iot.Hub.Service/src/IotHubServiceClient.cs
@@ -61,6 +61,8 @@ namespace Azure.Iot.Hub.Service
         /// </summary>
         public virtual JobsClient Jobs { get; private set; }
 
+        public virtual QueryClient Query { get; private set; }
+
         /// <summary>
         /// Initializes a new instance of the <see cref="IotHubServiceClient"/> class.
         /// </summary>
@@ -146,8 +148,13 @@ namespace Azure.Iot.Hub.Service
             _statisticsRestClient = new StatisticsRestClient(_clientDiagnostics, _httpPipeline, credential.Endpoint, options.GetVersionString());
             _configurationRestClient = new ConfigurationRestClient(_clientDiagnostics, _httpPipeline, credential.Endpoint, options.GetVersionString());
 
-            Devices = new DevicesClient(_devicesRestClient, _queryRestClient);
-            Modules = new ModulesClient(_devicesRestClient, _modulesRestClient, _queryRestClient);
+            // Note that the devices and modules subclient take a reference to the Query convenience layer client. This
+            // is because they each expose a helper function that uses the query client for listing twins. By passing in
+            // the convenience layer query client rather than the protocol layer query client, we minimize rewriting the
+            // same pagination logic that now exists only in the query convenience layer client.
+            Query = new QueryClient(_queryRestClient);
+            Devices = new DevicesClient(_devicesRestClient, Query);
+            Modules = new ModulesClient(_devicesRestClient, _modulesRestClient, Query);
             Statistics = new StatisticsClient(_statisticsRestClient);
             Configurations = new ConfigurationsClient(_configurationRestClient);
