+++ sdk/iot/Azure.Iot.Hub.Service/src/ModulesClient.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System.Collections.Generic;
-using System.Globalization;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -17,12 +16,11 @@ namespace Azure.Iot.Hub.Service
     /// </summary>
     public class ModulesClient
     {
-        private const string ContinuationTokenHeader = "x-ms-continuation";
         private const string HubModuleQuery = "select * from devices.modules";
 
         private readonly DevicesRestClient _devicesRestClient;
         private readonly ModulesRestClient _modulesRestClient;
-        private readonly QueryRestClient _queryRestClient;
+        private readonly QueryClient _queryClient;
 
         /// <summary>
         /// Initializes a new instance of ModulesClient.
@@ -35,17 +33,17 @@ namespace Azure.Iot.Hub.Service
         /// Initializes a new instance of DevicesClient.
         /// <param name="devicesRestClient"> The REST client to perform bulk operations on the module. </param>
         /// <param name="modulesRestClient"> The REST client to perform module and module twin operations. </param>
-        /// <param name="queryRestClient"> The REST client to perform query operations for the device. </param>
+        /// <param name="queryClient"> The convenience layer query client to perform query operations for the device. </param>
         /// </summary>
-        internal ModulesClient(DevicesRestClient devicesRestClient, ModulesRestClient modulesRestClient, QueryRestClient queryRestClient)
+        internal ModulesClient(DevicesRestClient devicesRestClient, ModulesRestClient modulesRestClient, QueryClient queryClient)
         {
             Argument.AssertNotNull(devicesRestClient, nameof(devicesRestClient));
             Argument.AssertNotNull(modulesRestClient, nameof(modulesRestClient));
-            Argument.AssertNotNull(queryRestClient, nameof(queryRestClient));
+            Argument.AssertNotNull(queryClient, nameof(queryClient));
 
             _devicesRestClient = devicesRestClient;
             _modulesRestClient = modulesRestClient;
-            _queryRestClient = queryRestClient;
+            _queryClient = queryClient;
         }
 
         /// <summary>
@@ -365,80 +363,29 @@ namespace Azure.Iot.Hub.Service
         /// <summary>
         /// List a set of module twins.
         /// </summary>
+        /// <remarks>
+        /// This service request returns the full set of module twins. To get a subset of module twins, you can use the <see cref="QueryClient.QueryAsync(string, int?, CancellationToken)">query API</see> that this method uses but with additional qualifiers for selection.
+        /// </remarks>
         /// <param name="pageSize">The size of each page to be retrieved from the service. Service may override this size.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>A pageable set of module twins <see cref="AsyncPageable{T}"/>.</returns>
         public virtual AsyncPageable<TwinData> GetTwinsAsync(int? pageSize = null, CancellationToken cancellationToken = default)
         {
-            async Task<Page<TwinData>> FirstPageFunc(int? pageSizeHint)
-            {
-                var querySpecification = new QuerySpecification
-                {
-                    Query = HubModuleQuery
-                };
-
-                Response<IReadOnlyList<TwinData>> response =
-                    await _queryRestClient.GetTwinsAsync(querySpecification, null, pageSizeHint?.ToString(CultureInfo.InvariantCulture), cancellationToken).ConfigureAwait(false);
-
-                response.GetRawResponse().Headers.TryGetValue(ContinuationTokenHeader, out string continuationToken);
-
-                return Page.FromValues(response.Value, continuationToken, response.GetRawResponse());
-            }
-
-            async Task<Page<TwinData>> NextPageFunc(string nextLink, int? pageSizeHint)
-            {
-                var querySpecification = new QuerySpecification();
-
-                Response<IReadOnlyList<TwinData>> response =
-                    await _queryRestClient.GetTwinsAsync(querySpecification, nextLink, pageSizeHint?.ToString(CultureInfo.InvariantCulture), cancellationToken).ConfigureAwait(false);
-
-                response.GetRawResponse().Headers.TryGetValue(ContinuationTokenHeader, out string continuationToken);
-                return Page.FromValues(response.Value, continuationToken, response.GetRawResponse());
-            }
-
-            return PageableHelpers.CreateAsyncEnumerable(FirstPageFunc, NextPageFunc, pageSize);
+            return _queryClient.QueryAsync(HubModuleQuery, pageSize, cancellationToken);
         }
 
         /// <summary>
         /// List a set of module twins.
         /// </summary>
+        /// <remarks>
+        /// This service request returns the full set of module twins. To get a subset of module twins, you can use the <see cref="QueryClient.Query(string, int?, CancellationToken)">query API</see> that this method uses but with additional qualifiers for selection.
+        /// </remarks>
         /// <param name="pageSize">The size of each page to be retrieved from the service. Service may override this size.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>A pageable set of module twins <see cref="Pageable{T}"/>.</returns>
         public virtual Pageable<TwinData> GetTwins(int? pageSize = null, CancellationToken cancellationToken = default)
         {
-            Page<TwinData> FirstPageFunc(int? pageSizeHint)
-            {
-                var querySpecification = new QuerySpecification
-                {
-                    Query = HubModuleQuery
-                };
-
-                Response<IReadOnlyList<TwinData>> response = _queryRestClient.GetTwins(
-                    querySpecification,
-                    null,
-                    pageSizeHint?.ToString(CultureInfo.InvariantCulture),
-                    cancellationToken);
-
-                response.GetRawResponse().Headers.TryGetValue(ContinuationTokenHeader, out string continuationToken);
-
-                return Page.FromValues(response.Value, continuationToken, response.GetRawResponse());
-            }
-
-            Page<TwinData> NextPageFunc(string nextLink, int? pageSizeHint)
-            {
-                var querySpecification = new QuerySpecification();
-                Response<IReadOnlyList<TwinData>> response = _queryRestClient.GetTwins(
-                    querySpecification,
-                    nextLink,
-                    pageSizeHint?.ToString(CultureInfo.InvariantCulture),
-                    cancellationToken);
-
-                response.GetRawResponse().Headers.TryGetValue(ContinuationTokenHeader, out string continuationToken);
-                return Page.FromValues(response.Value, continuationToken, response.GetRawResponse());
-            }
-
-            return PageableHelpers.CreateEnumerable(FirstPageFunc, NextPageFunc, pageSize);
+            return _queryClient.Query(HubModuleQuery, pageSize, cancellationToken);
         }
 
         /// <summary>
